Move WriteHandlerLookupFailureTest to its own class
It shouldn't be part of the clustered event bus test suite.
